Ošetření mazaní kategorií

+++ app/model/CategoryManager.php
@@ -2,10 +2,11 @@
 
 namespace App\model;
 
+
 /** Author: Dalibor Kyjovský */
 
 use App\repository\CategoryRepository;
-use App\model\UserManager;
+use Nextras\Dbal\Connection;
 use Exception;
 
 class CategoryManager
@@ -14,15 +15,36 @@ class CategoryManager
     /** @var CategoryRepository */
     private $categoryRepository;
 
-    public function __construct(CategoryRepository $categoryRepository)
+    /** @var $connection */
+    public $connection;
+
+    public function __construct(CategoryRepository $categoryRepository, Connection $connection)
     {
         $this->categoryRepository = $categoryRepository;
+        $this->connection = $connection;
     }
 
 
     public function delete($id)
     {
-        $this->categoryRepository->deleteCategoryByUserId($id);
+        $categories_id = $this->connection->query("SELECT `expenses`.`categories_id` FROM `expenses`")->fetchall();
+        $test = 0;
+
+
+        foreach ($categories_id as $used_id) {
+            if ($used_id->categories_id == $id) {
+                $test++;
+            }
+
+        }
+        
+        if($test == 0)
+        {
+            $this->categoryRepository->deleteCategoryByUserId($id);
+        }
+        else {
+            return 1;
+        }
 
     }
 }
